+++ lib/audited-rspec.rb
@@ -1,4 +1,6 @@
-require 'audited/rspec_matchers'
+# frozen_string_literal: true
+
+require "audited/rspec_matchers"
 module RSpec::Matchers
   include Audited::RspecMatchers
 end