+++ lib/generators/audited/install_generator.rb
@@ -1,9 +1,11 @@
-require 'rails/generators'
-require 'rails/generators/migration'
-require 'active_record'
-require 'rails/generators/active_record'
-require 'generators/audited/migration'
-require 'generators/audited/migration_helper'
+# frozen_string_literal: true
+
+require "rails/generators"
+require "rails/generators/migration"
+require "active_record"
+require "rails/generators/active_record"
+require "generators/audited/migration"
+require "generators/audited/migration_helper"
 
 module Audited
   module Generators
@@ -18,7 +20,7 @@ module Audited
       source_root File.expand_path("../templates", __FILE__)
 
       def copy_migration
-        migration_template 'install.rb', 'db/migrate/install_audited.rb'
+        migration_template "install.rb", "db/migrate/install_audited.rb"
       end
     end
   end