+++ lib/generators/audited/migration_helper.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 module Audited
   module Generators
     module MigrationHelper
       def migration_parent
-        Rails::VERSION::MAJOR == 4 ? 'ActiveRecord::Migration' : "ActiveRecord::Migration[#{ActiveRecord::Migration.current_version}]"
+        "ActiveRecord::Migration[#{ActiveRecord::Migration.current_version}]"
       end
     end
   end