+++ spec/rails_app/config/environments/test.rb
@@ -15,14 +15,14 @@ RailsApp::Application.configure do
   # Configure static file server for tests with Cache-Control for performance.
   if config.respond_to?(:public_file_server)
     config.public_file_server.enabled = true
-    config.public_file_server.headers = { 'Cache-Control' => 'public, max-age=3600' }
+    config.public_file_server.headers = {"Cache-Control" => "public, max-age=3600"}
   else
-    config.static_cache_control = 'public, max-age=3600'
-    config.serve_static_files   = true
+    config.static_cache_control = "public, max-age=3600"
+    config.serve_static_files = true
   end
 
   # Show full error reports and disable caching.
-  config.consider_all_requests_local       = true
+  config.consider_all_requests_local = true
   # config.action_controller.perform_caching = false
 
   # Raise exceptions instead of rendering exception templates.
@@ -34,7 +34,7 @@ RailsApp::Application.configure do
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
-  config.action_mailer.delivery_method = :test
+  # config.action_mailer.delivery_method = :test
 
   # Randomize the order test cases are executed.
   config.active_support.test_order = :random