+++ test/db/version_2.rb
@@ -12,7 +12,7 @@ ActiveRecord::Schema.define do
     t.column :created_at, :datetime
   end
 
-  add_index :audits, [:auditable_id, :auditable_type], name: 'auditable_index'
-  add_index :audits, [:user_id, :user_type], name: 'user_index'
+  add_index :audits, [:auditable_id, :auditable_type], name: "auditable_index"
+  add_index :audits, [:user_id, :user_type], name: "user_index"
   add_index :audits, :created_at
 end