+++ test/db/version_6.rb
@@ -14,4 +14,6 @@ ActiveRecord::Schema.define do
     t.column :associated_id, :integer
     t.column :associated_type, :string
   end
+
+  add_index :audits, [:auditable_type, :auditable_id], name: "auditable_index"
 end